Refactor: Update ID types in DAOs and entities
This commit updates the type of ID fields in various DAOs (Data Access Objects) and entities from `Int` to `Long`. This change provides greater flexibility and scalability for handling larger datasets.

Specifically, the following changes were made:

- **ContactTierDao:**
    - `getTierById` and `getTierWithContactsById` now accept `Long` instead of `Int` for the ID parameter.
- **ContactTagCrossRefDao:**
    - `removeAllTagsFromContact` now accepts `Long` instead of `Int` for the `contactId` parameter.
- **ContactDao:**
    - `getContactsByTierId` and `getContactById` now accept `Long` instead of `Int` for the ID parameter.
    - Added `getContactsByTagId` method to retrieve contacts associated with a specific tag ID (using `Long`).
- **ContactEntity:**
    - Added `@ColumnInfo(index = true)` to `tierId` for improved query performance.
- **ContactTagCrossRef:**
    - Added `@ColumnInfo(index = true)` to `tagId` for improved query performance.
- **ContactDataMappers:**
    - Added `toListEntity` and `toListModel` extension functions for converting lists of DTOs/entities to entities/models.

Additionally, a new run configuration for `ContactTagCrossRefDaoTest` was added to the `.idea/deploymentTargetSelector.xml` file.

+++ app/src/main/java/com/hollowvyn/kneatr/data/local/dao/ContactTagCrossRefDao.kt
@@ -16,5 +16,5 @@ interface ContactTagCrossRefDao {
     suspend fun removeTagFromContact(crossRef: ContactTagCrossRef)
 
     @Query("DELETE FROM ContactTagCrossRef WHERE contactId = :contactId")
-    suspend fun removeAllTagsFromContact(contactId: Int)
+    suspend fun removeAllTagsFromContact(contactId: Long)
 }
